refactor: 패키지 의존성 분리
