+++ src/main/java/kitchenpos/table/application/TableGroupService.java
@@ -1,6 +1,6 @@
 package kitchenpos.table.application;
 
-import kitchenpos.order.repository.OrderTableRepository;
+import kitchenpos.table.repository.OrderTableRepository;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.TableGroup;
 import kitchenpos.table.presentation.dto.TableGroupCreateRequest;
