+++ src/test/java/kitchenpos/application/OrderTableServiceTest.java
@@ -4,11 +4,11 @@ import kitchenpos.Product.repository.ProductRepository;
 import kitchenpos.menu.repository.MenuProductRepository;
 import kitchenpos.menu.repository.MenuRepository;
 import kitchenpos.menugroup.repository.MenuGroupRepository;
-import kitchenpos.order.presentation.dto.OrderTableCreateRequest;
-import kitchenpos.order.presentation.dto.OrderTableUpdateEmptyRequest;
-import kitchenpos.order.presentation.dto.OrderTableUpdateNumberOfGuestsRequest;
+import kitchenpos.table.presentation.dto.OrderTableCreateRequest;
+import kitchenpos.table.presentation.dto.OrderTableUpdateEmptyRequest;
+import kitchenpos.table.presentation.dto.OrderTableUpdateNumberOfGuestsRequest;
 import kitchenpos.order.repository.OrderRepository;
-import kitchenpos.order.repository.OrderTableRepository;
+import kitchenpos.table.repository.OrderTableRepository;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.repository.TableGroupRepository;
 import org.junit.jupiter.api.Test;
